Fix dependency pour le JPA

+++ src/main/java/fr/isima/tp1/repository/RuleRepository.java
@@ -7,6 +7,6 @@ import org.springframework.stereotype.Repository;
 
 
 @Repository
-public interface RuleRepository extends JpaRepository<Rule, Long> {
+public interface RuleRepository extends JpaRepository<Rule, Integer> {
 
 }
